2 composants arrox-left et add-artist
