+++ packages/frontend/web/src/pages/create-album-menu.tsx
@@ -1,29 +1,17 @@
-import { useNavigate } from 'react-router';
-import { Link } from 'react-router-dom';
-
+import AddArtist from '@/components/add-artist';
 import AddButton from '@/components/add-button';
 import AddSingle from '@/components/add-single';
+import { ArrowLeft } from '@/components/arrow-left';
 import ChooseName from '@/components/choose-name';
 import RemoveSingle from '@/components/remove-single';
 import VerifyButton from '@/components/verify-button';
 
 export default function CreateAlbumMenu() {
-  const navigate = useNavigate();
   return (
     <div className='bg-primary flex min-h-screen flex-col items-center px-4 py-6'>
       <div className='mb-4 flex w-full items-center justify-between'>
-        <button
-          type='button'
-          className='text-secondary hover:text-orange-500'
-          onClick={async () => {
-            await navigate(-1);
-          }}
-        >
-          <img
-            src='/assets/arrow-left.png'
-            alt='arrow left'
-            className='W-10 h-10'
-          />
+        <button type='button'>
+          <ArrowLeft />
         </button>
         <h1 className='text-secondary text-center text-2xl font-bold'>
           {'RECORDING A NEW ALBUM'}
@@ -31,12 +19,8 @@ export default function CreateAlbumMenu() {
         <div className='h-6 w-6' />
       </div>
       <img className='h-22 w-22' src='/assets/album.png' alt='' />
-      <div className='mt-8 flex flex-col items-center justify-center'>
-        <Link to='/my-artists'>
-          <AddButton>{'+'}</AddButton>
-        </Link>
-
-        <h2 className='text-secondary pt-1 text-xl'>{'Choose an artist'}</h2>
+      <div>
+        <AddArtist />
       </div>
       <div className='flex flex-col items-center justify-center'>
         <ChooseName
