+++ packages/frontend/web/src/pages/create-single-menu.tsx
@@ -1,27 +1,15 @@
-import { useNavigate } from 'react-router';
-import { Link } from 'react-router-dom';
-
+import AddArtist from '@/components/add-artist';
 import AddButton from '@/components/add-button';
+import { ArrowLeft } from '@/components/arrow-left';
 import ChooseName from '@/components/choose-name';
 import VerifyButton from '@/components/verify-button';
 
 export default function CreateSingleMenu() {
-  const navigate = useNavigate();
   return (
     <div className='bg-primary flex min-h-screen flex-col items-center px-4 py-6'>
       <div className='mb-4 flex w-full items-center justify-between'>
-        <button
-          type='button'
-          className='text-secondary hover:text-orange-500'
-          onClick={async () => {
-            await navigate(-1);
-          }}
-        >
-          <img
-            src='/assets/arrow-left.png'
-            alt='arrow left'
-            className='W-10 h-10'
-          />
+        <button type='button'>
+          <ArrowLeft />
         </button>
         <h1 className='text-secondary text-center text-2xl font-bold'>
           {'RECORDING A NEW SINGLE'}
@@ -30,12 +18,8 @@ export default function CreateSingleMenu() {
       </div>
       <img className='h-18 w-18' src='/assets/music-note.png' alt='' />
 
-      <div className='mt-8 flex flex-col items-center justify-center'>
-        <Link to='/my-artists'>
-          <AddButton>{'+'}</AddButton>
-        </Link>
-
-        <h2 className='text-secondary pt-1 text-xl'>{'Choose an artist'}</h2>
+      <div>
+        <AddArtist />
       </div>
       <div className='flex flex-col items-center justify-center'>
         <ChooseName
